stash add: interface type handling component
