+++ packages/app-schema-editor/by-selection-state/selection-variant-master.tsx
@@ -12,17 +12,15 @@ import {
   jsxViewExampleBuilder,
 } from "../interface-code-builder";
 import { nameit, NameCases } from "@coli.codes/naming";
-import { Interface } from "@code-ui/interface";
 import {
   InterfaceAttr,
   InterfaceProps,
   InterfaceTypeOption,
 } from "@code-ui/interface/dist/lib/type";
+import { SelectionCodeUiInterface } from "./selectoin-code-ui-interface";
 
 export default function (props: { node: nodes.light.IReflectNodeReference }) {
   const master = props.node;
-  const regxNumberType = /(number)/;
-  const regxBooleanType = /(boolean)/;
 
   const parser = new VariantPropertyParser(master);
   const data_of_properties = parser.getData(master);
@@ -30,37 +28,6 @@ export default function (props: { node: nodes.light.IReflectNodeReference }) {
     case: NameCases.pascal,
   }).name;
 
-  function typeToString(_data: FigmaVariantPropertyCompatType): any {
-    if (typeof _data === "symbol") {
-      const isNumberType = regxNumberType.test(_data.toString());
-      const isBooleanType = regxBooleanType.test(_data.toString());
-
-      if (isNumberType) {
-        return "number";
-      } else if (isBooleanType) {
-        // will be change to boolean
-        // it just for interface package typo missing
-        return "bool";
-      }
-    } else if (_data.type) {
-      return _data.type;
-    } else {
-      return "any";
-    }
-  }
-
-  const interfaceAttrs: InterfaceAttr[] = parser.properties.map((d) => {
-    const attrTypes = typeToString(d.type);
-    const _contorl: InterfaceTypeOption = {
-      name: attrTypes,
-      value: attrTypes,
-      description: `type is ${attrTypes}`,
-    };
-    return {
-      label: d.key,
-      contorls: [_contorl],
-    };
-  });
   const viewName = nameit(master.parent.name, {
     case: NameCases.pascal,
   }).name;
@@ -69,11 +36,9 @@ export default function (props: { node: nodes.light.IReflectNodeReference }) {
     <>
       <h6>variant</h6>
 
-      <Interface
-        lang={"js"}
-        theme={"monokai"}
+      <SelectionCodeUiInterface
+        properties={parser.properties}
         interfaceName={interfaceName}
-        attrs={interfaceAttrs}
         onChange={() => {}}
       />
       <CodeBox
